Added Short and Long Idle Animations to the Player Spritesheet
Both the short and idle animations exist for upwards, downwards, and sideways facing directions of the player character.

The short animation should be used any time the player is not pressing a key when facing a certain direction. The long animation should only be used when the player has not performed any action within a larger span of time.

*The three sprites at the bottom of the spritesheet will be for when the player takes damage, but are not completed.