Reworked Player Punch to be on a Separate Spritesheet
The third and fifth sequence of the punch animation has been changed for all player directions on player_spritesheet.png. The actual arm and fist are now removed from the spritesheet and in its place is a separate spritesheet named spritesheet_playerattack.png.

When the third and fifth sequence in the spritesheet occurs within the game (the actual punch), the new sprite sheet should be placed on a separate layer to create the player's attacks which will have their own hitboxes.